Using matrices properly now - looks WAY better
Matrices handle the transforms properly now. Still doesn't render
perfectly, but a little bit of font reading and we're sorted. Once we've
interpreted the fonts and done proper word and character spacing then we
can trust where the glyphs are in memory, and make better guesses about
where they line up in a spreadsheet